Price and shopping basket fully implemented
Now the price is completly retrive from the database using the
Query builder of CodeIgniter. As this is done connecting to the
databse you will see these changes in the controllers.

Now, in the shopping basket, in the renting page you will
see the complete cost with all the products as well as the
time amount of each one of them. The complete cost calculation
is done in the Members.php controller.