Process payment funtion added
In the members controller there's a funtion called process_products_payment,
right now there's no payment handling like paypal or payeer, so right now
all this function does is sent notification to the product's owner and product's lessee.

It also clears the shopping basket and upload the rentings to the database.

I also updated the basket dropdown list in nav, now it links to the full shopping basket
page, that way the link to that page is not in the user's dropdown list